modified:   Tropic-Island/Domini/Chess/Chess/Chess.cpp 	modified:   Tropic-Island/Domini/Chess/Chess/Game.cpp 	modified:   Tropic-Island/Domini/Chess/Chess/Game.h 	modified:   Tropic-Island/Domini/Chess/Chess/WindowsSDLApi.cpp 	modified:   Tropic-Island/Domini/Chess/Chess/WindowsSDLApi.h

+++ Tropic-Island/Domini/Chess/Chess/WindowsSDLApi.h
@@ -12,8 +12,8 @@
     #define _POINT_
     typedef struct POINT
     {
-        float x;
-        float y;
+        int x;
+        int y;
     };
     #endif
 #endif
@@ -32,7 +32,6 @@ public:
     POINT GetButtonDownCoords(SDL_Event event_);
     bool IsExit(SDL_Event event_);
     byte GetPressedKey(SDL_Event event_);
-
     bool getdone();
     void setdone(bool state);
 	int pressbutton;
